@@ -1,23 +1,27 @@
 {
-  "name": "e2e",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@types/cypress": "^1.1.3",
-    "dedent": "^0.7.0",
-    "typescript": "^4.7.2"
-  },
-  "scripts": {
-    "start": "ts-node process-manager/start.ts"
-  },
-  "devDependencies": {
-    "@types/node": "^17.0.38",
-    "body-parser": "^1.20.0",
-    "cors": "^2.8.5",
-    "cypress-promise": "^1.1.0",
-    "express": "^4.18.1",
-    "isomorphic-fetch": "^3.0.0",
-    "ts-node": "^10.8.0"
-  }
+    "name": "e2e",
+    "version": "1.0.0",
+    "main": "index.js",
+    "license": "MIT",
+    "dependencies": {
+        "@types/cypress": "^1.1.3",
+        "dedent": "^0.7.0",
+        "typescript": "^4.9.3"
+    },
+    "scripts": {
+        "start": "ts-node process-manager/start.ts",
+        "cy:open": "cypress open",
+        "cy:run": "cypress run",
+        "cy:run-head": "cypress run --headed --no-exit"
+    },
+    "devDependencies": {
+        "@types/node": "^18.11.10",
+        "body-parser": "^1.20.1",
+        "cors": "^2.8.5",
+        "cypress": "^11.2.0",
+        "cypress-promise": "^1.1.0",
+        "express": "^4.18.2",
+        "isomorphic-fetch": "^3.0.0",
+        "ts-node": "^10.9.1"
+    }
 }
