@@ -1,8 +1,8 @@
-import {start} from './index'
+import { start } from "./index";
 
-console.log('starting fiftyone e2e tests...')
+console.log("starting fiftyone e2e tests...");
 
-start().catch(e => {
-  console.error(e)
-  process.exit(1)
-})
+start().catch((e) => {
+  console.error(e);
+  process.exit(1);
+});
